Misc. changes (#253)
* add link to tools docs

* add beta tooltip

* fix tooltip for button

* see output from dist replacement

* echo

* Update release version to dev-testing

* try publishing a docker runtime image

* separate login from build

* Update release version to dev-testing

* tag with the repo

* Update release version to dev-testing

* cloudera registry in ghcr.io?

* Update release version to dev-testing

* add a repo-assembly.json file

* try setting the user to cdsw

* Update release version to dev-testing

* set some directory perms for cdsw

* Update release version to dev-testing

* update to point at the dev-testing

* try forcing python 3.12 for uv

* Update release version to dev-testing-2

* allow more pythons

* Update release version to dev-testing-2

* fix uv.lock

* Update release version to dev-testing-2

* set a VIRTUAL_ENV to the possibly pre-installed venv

* try the UV_PROJECT_ENVIRONMENT

* put the dependencies in the docker image

* Update release version to dev-testing-3

* install node in the docker container

* update to v4

* Update release version to dev-testing-4

* don't install node if it's already installed

* update a bunch of deps

* tiny comment update

* move the runtime build/publish to a separate GHA

* fix an old mypy issue

---------

+++ scripts/install_easyocr_model.sh
@@ -44,6 +44,13 @@ LATIN_G2_URL=https://github.com/cloudera/CML_AMP_RAG_Studio/releases/download/mo
 mkdir -p .EasyOCR/model
 cd .EasyOCR/model
 
-wget --no-verbose -O craft_mlt_25k.pth ${CRAFT_MLT_MODEL_URL}
-wget --no-verbose -O latin_g2.pth ${LATIN_G2_URL}
+if [ -f /app/craft_mlt_25k.pth ]; then
+    cp /app/craft_mlt_25k.pth craft_mlt_25k.pth
+    cp /app/latin_g2.pth latin_g2.pth
+else
+    wget --no-verbose -O craft_mlt_25k.pth ${CRAFT_MLT_MODEL_URL}
+    wget --no-verbose -O latin_g2.pth ${LATIN_G2_URL}
+fi
+
+
